Refactoring TTrapSender.
